+++ silva/src/components/workflow/mod.rs
@@ -1,12 +1,12 @@
 pub mod home;
-pub mod job;
+pub mod job_folder;
 pub mod manager;
 pub mod param_source;
 pub mod params_editor;
 pub mod workflow_folder;
 
 pub use home::{WorkflowHome, WorkflowHomeError};
-pub use job::{Job, JobError, JobScanner};
+pub use job_folder::{JobError, JobFolder, JobScanner};
 pub use manager::{WorkflowError, WorkflowManager};
 pub use param_source::{JobParamSource, ParamSource, WorkflowParamSource};
 pub use params_editor::ParamsEditorState;